Compiles

+++ include/stack.h
@@ -11,11 +11,11 @@ typedef struct Stack
 {
   uintptr_t sp;
   uintptr_t fp;
-  unsigned int data_size;
+  size_t data_size;
   uintptr_t *data;
 } Stack;
 
-void stack_init (Stack *s_ptr, unsigned int count);
+void stack_init (Stack *s_ptr, size_t count);
 void stack_free (Stack *s_ptr);
 void stack_push (Stack *s_ptr, void *value);
 void *stack_pop (Stack *s_ptr);
